use single module to handle python pure implementation of numba one
