+++ napari/layers/shapes/_shapes_models/shape.py
@@ -6,6 +6,9 @@ from functools import cached_property
 import numpy as np
 import numpy.typing as npt
 
+from napari.layers.shapes._accelerated_triangulate_wrap import (
+    remove_path_duplicates,
+)
 from napari.layers.shapes._shapes_utils import (
     is_collinear,
     path_to_mask,
@@ -28,33 +31,6 @@ except ImportError:
     triangulate_path_edge_numpy = None
     triangulate_polygon_numpy_li = None
     triangulate_polygon_with_edge_numpy_li = None
-
-
-def _remove_path_duplicates_np(data: np.ndarray, closed: bool):
-    # We add the first data point at the end to get the same length bool
-    # array as the data, and also to work on closed shapes; the last value
-    # in the diff array compares the last and first vertex.
-    diff = np.diff(np.append(data, data[0:1, :], axis=0), axis=0)
-    dup = np.all(diff == 0, axis=1)
-    # if the shape is closed, check whether the first vertex is the same
-    # as the last vertex, and count it as a duplicate if so
-    if closed and dup[-1]:
-        dup[0] = True
-    # we allow repeated nodes at the end for the lasso tool, which
-    # for an instant needs both the last placed point and the point at the
-    # cursor to be the same; if the lasso implementation becomes cleaner,
-    # remove this hardcoding
-    dup[-2:] = False
-    indices = np.arange(data.shape[0])
-    return data[indices[~dup]]
-
-
-try:
-    from napari.layers.shapes._accelerated_triangulate import (
-        remove_path_duplicates,
-    )
-except ImportError:
-    remove_path_duplicates = _remove_path_duplicates_np
 
 
 class Shape(ABC):
