+++ napari/layers/shapes/_shapes_utils.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import itertools
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -7,7 +8,9 @@ from skimage.draw import line, polygon2mask
 from vispy.geometry import PolygonData
 from vispy.visuals.tube import _frenet_frames
 
-from napari.layers.utils.layer_utils import segment_normal
+from napari.layers.shapes._accelerated_triangulate_wrap import (
+    generate_2D_edge_meshes_py,
+)
 from napari.utils.translations import trans
 
 if TYPE_CHECKING:
@@ -18,16 +21,6 @@ try:
     from triangle import triangulate
 except ModuleNotFoundError:
     triangulate = None
-
-
-try:
-    from napari.layers.shapes._accelerated_triangulate import (
-        create_box_from_bounding as acc_create_box_from_bounding,
-        generate_2D_edge_meshes as acc_generate_2D_edge_meshes,
-    )
-except ImportError:
-    acc_generate_2D_edge_meshes = None
-    acc_create_box_from_bounding = None
 
 
 def _is_convex(poly: npt.NDArray) -> bool:
@@ -269,12 +262,9 @@ def on_segment(p, q, r):
     on : bool
         Bool indicating if q is on segment from p to r
     """
-    if (
-        q[0] <= max(p[0], r[0])
-        and q[0] >= min(p[0], r[0])
-        and q[1] <= max(p[1], r[1])
-        and q[1] >= min(p[1], r[1])
-    ):
+    if max(p[0], r[0]) >= q[0] >= min(p[0], r[0]) and max(p[1], r[1]) >= q[
+        1
+    ] >= min(p[1], r[1]):
         on = True
     else:
         on = False
@@ -387,41 +377,6 @@ def point_to_lines(point, lines):
     return index, location
 
 
-def create_box_from_bounding(bounding_box: npt.NDArray) -> npt.NDArray:
-    """Creates the axis aligned interaction box of a bounding box
-
-    Parameters
-    ----------
-    bounding_box : np.ndarray
-        2x2 array of the bounding box. The first row is the minimum values and
-        the second row is the maximum values
-
-    Returns
-    -------
-    box : np.ndarray
-        9x2 array of vertices of the interaction box. The first 8 points are
-        the corners and midpoints of the box in clockwise order starting in the
-        upper-left corner. The last point is the center of the box
-    """
-    tl = bounding_box[(0, 0), (0, 1)]
-    br = bounding_box[(1, 1), (0, 1)]
-    tr = bounding_box[(1, 0), (0, 1)]
-    bl = bounding_box[(0, 1), (0, 1)]
-    return np.array(
-        [
-            tl,
-            (tl + tr) / 2,
-            tr,
-            (tr + br) / 2,
-            br,
-            (br + bl) / 2,
-            bl,
-            (bl + tl) / 2,
-            (tl + br) / 2,
-        ]
-    )
-
-
 def create_box(data: npt.NDArray) -> npt.NDArray:
     """Creates the axis aligned interaction box of a list of points
 
@@ -476,7 +431,7 @@ def rectangle_to_box(data: npt.NDArray) -> npt.NDArray:
         the corners and midpoints of the box in clockwise order starting in the
         upper-left corner. The last point is the center of the box
     """
-    if not data.shape[0] == 4:
+    if data.shape[0] != 4:
         raise ValueError(
             trans._(
                 'Data shape does not match expected `[4, D]` shape specifying corners for the rectangle',
@@ -586,7 +541,7 @@ def triangulate_ellipse(
 
 
     """
-    if not corners.shape[0] == 4:
+    if corners.shape[0] != 4:
         raise ValueError(
             trans._(
                 'Data shape does not match expected `[4, D]` shape specifying corners for the ellipse',
@@ -713,7 +668,7 @@ def triangulate_edge(
         clean_path = path
 
     if clean_path.shape[-1] == 2:
-        centers, offsets, triangles = _generate_2D_edge_meshes(
+        centers, offsets, triangles = generate_2D_edge_meshes_py(
             np.asarray(clean_path, dtype=np.float32), closed=closed
         )
     else:
@@ -722,179 +677,6 @@ def triangulate_edge(
         )
 
     # offsets[2,1] = -0.5
-    return centers, offsets, triangles
-
-
-def _mirror_point(x, y):
-    return 2 * y - x
-
-
-def _sign_nonzero(x):
-    y = np.sign(x).astype(int)
-    y[y == 0] = 1
-    return y
-
-
-def _sign_cross(x, y):
-    """sign of cross product (faster for 2d)"""
-    if x.shape[1] == y.shape[1] == 2:
-        return _sign_nonzero(x[:, 0] * y[:, 1] - x[:, 1] * y[:, 0])
-    if x.shape[1] == y.shape[1] == 3:
-        return _sign_nonzero(np.cross(x, y))
-
-    raise ValueError(x.shape[1], y.shape[1])
-
-
-def generate_2D_edge_meshes(path, closed=False, limit=3, bevel=False):
-    """Determines the triangulation of a path in 2D. The resulting `offsets`
-    can be multiplied by a `width` scalar and be added to the resulting
-    `centers` to generate the vertices of the triangles for the triangulation,
-    i.e. `vertices = centers + width*offsets`. Using the `centers` and
-    `offsets` representation thus allows for the computed triangulation to be
-    independent of the line width.
-
-    Parameters
-    ----------
-    path : np.ndarray
-        Nx2 or Nx3 array of central coordinates of path to be triangulated
-    closed : bool
-        Bool which determines if the path is closed or not
-    limit : float
-        Miter limit which determines when to switch from a miter join to a
-        bevel join
-    bevel : bool
-        Bool which if True causes a bevel join to always be used. If False
-        a bevel join will only be used when the miter limit is exceeded
-
-    Returns
-    -------
-    centers : np.ndarray
-        Mx2 or Mx3 array central coordinates of path triangles.
-    offsets : np.ndarray
-        Mx2 or Mx3 array of the offsets to the central coordinates that need to
-        be scaled by the line width and then added to the centers to
-        generate the actual vertices of the triangulation
-    triangles : np.ndarray
-        Px3 array of the indices of the vertices that will form the
-        triangles of the triangulation
-    """
-
-    path = np.asarray(path, dtype=float)
-
-    # add first vertex to the end if closed
-    if closed:
-        path = np.concatenate((path, [path[0]]))
-
-    # extend path by adding a vertex at beginning and end
-    # to get the mean normals correct
-    if closed:
-        _ext_point1 = path[-2]
-        _ext_point2 = path[1]
-    else:
-        _ext_point1 = _mirror_point(path[1], path[0])
-        _ext_point2 = _mirror_point(path[-2], path[-1])
-
-    full_path = np.concatenate(([_ext_point1], path, [_ext_point2]), axis=0)
-
-    # full_normals[:-1], full_normals[1:] are normals left and right of each path vertex
-    full_normals = segment_normal(full_path[:-1], full_path[1:])
-
-    # miters per vertex are the average of left and right normals
-    miters = 0.5 * (full_normals[:-1] + full_normals[1:])
-
-    # scale miters such that their dot product with normals is 1
-    _mf_dot = np.expand_dims(
-        np.einsum('ij,ij->i', miters, full_normals[:-1]), -1
-    )
-
-    miters = np.divide(
-        miters,
-        _mf_dot,
-        where=np.abs(_mf_dot) > 1e-10,
-    )
-
-    miter_lengths_squared = (miters**2).sum(axis=1)
-
-    # miter_signs -> +1 if edges turn clockwise, -1 if anticlockwise
-    # used later to discern bevel positions
-    miter_signs = _sign_cross(full_normals[1:], full_normals[:-1])
-    miters = 0.5 * miters
-
-    # generate centers/offsets
-    centers = np.repeat(path, 2, axis=0)
-    offsets = np.repeat(miters, 2, axis=0)
-    offsets[::2] *= -1
-
-    triangles0 = np.tile(np.array([[0, 1, 3], [0, 3, 2]]), (len(path) - 1, 1))
-    triangles = triangles0 + 2 * np.repeat(
-        np.arange(len(path) - 1)[:, np.newaxis], 2, 0
-    )
-
-    # get vertex indices that are to be beveled
-    idx_bevel = np.where(
-        np.bitwise_or(bevel, miter_lengths_squared > (limit**2))
-    )[0]
-
-    if len(idx_bevel) > 0:
-        idx_offset = (miter_signs[idx_bevel] < 0).astype(int)
-
-        # outside and inside offsets are treated differently (only outside offsets get beveled)
-        # See drawing at:
-        # https://github.com/napari/napari/pull/6706#discussion_r1528790407
-        idx_bevel_outside = 2 * idx_bevel + idx_offset
-        idx_bevel_inside = 2 * idx_bevel + (1 - idx_offset)
-        sign_bevel = np.expand_dims(miter_signs[idx_bevel], -1)
-
-        # adjust offset of outer offset
-        offsets[idx_bevel_outside] = (
-            -0.5 * full_normals[:-1][idx_bevel] * sign_bevel
-        )
-        # adjust/normalize length of inner offset
-        offsets[idx_bevel_inside] /= np.sqrt(
-            miter_lengths_squared[idx_bevel, np.newaxis]
-        )
-
-        # special cases for the last vertex
-        _nonspecial = idx_bevel != len(path) - 1
-
-        idx_bevel = idx_bevel[_nonspecial]
-        idx_bevel_outside = idx_bevel_outside[_nonspecial]
-        sign_bevel = sign_bevel[_nonspecial]
-        idx_offset = idx_offset[_nonspecial]
-
-        # create new "right" bevel vertices to be added later
-        centers_bevel = path[idx_bevel]
-        offsets_bevel = -0.5 * full_normals[1:][idx_bevel] * sign_bevel
-
-        n_centers = len(centers)
-        # change vertices of triangles to the newly added right vertices
-        triangles[2 * idx_bevel, idx_offset] = len(centers) + np.arange(
-            len(idx_bevel)
-        )
-        triangles[2 * idx_bevel + (1 - idx_offset), idx_offset] = (
-            n_centers + np.arange(len(idx_bevel))
-        )
-
-        # add a new center/bevel triangle
-        triangles0 = np.tile(np.array([[0, 1, 2]]), (len(idx_bevel), 1))
-        triangles_bevel = np.array(
-            [
-                2 * idx_bevel + idx_offset,
-                2 * idx_bevel + (1 - idx_offset),
-                n_centers + np.arange(len(idx_bevel)),
-            ]
-        ).T
-        # add all new centers, offsets, and triangles
-        centers = np.concatenate([centers, centers_bevel])
-        offsets = np.concatenate([offsets, offsets_bevel])
-        triangles = np.concatenate([triangles, triangles_bevel])
-
-    # extracting vectors (~4x faster than np.moveaxis)
-    a, b, c = tuple((centers + offsets)[triangles][:, i] for i in range(3))
-    # flip negative oriented triangles
-    flip_idx = _sign_cross(b - a, c - a) < 0
-    triangles[flip_idx] = np.flip(triangles[flip_idx], axis=-1)
-
     return centers, offsets, triangles
 
 
@@ -928,7 +710,7 @@ def generate_tube_meshes(path, closed=False, tube_points=10):
     if closed and not np.array_equal(points[0], points[-1]):
         points = np.concatenate([points, [points[0]]], axis=0)
 
-    tangents, normals, binormals = _frenet_frames(points, closed)
+    _tangents, normals, binormals = _frenet_frames(points, closed)
 
     segments = len(points) - 1
 
@@ -949,18 +731,18 @@ def generate_tube_meshes(path, closed=False, tube_points=10):
 
     # construct the mesh
     indices = []
-    for i in range(segments):
-        for j in range(tube_points):
-            ip = (i + 1) % segments if closed else i + 1
-            jp = (j + 1) % tube_points
+    for i, j in itertools.product(range(segments), range(tube_points)):
+        ip = (i + 1) % segments if closed else i + 1
+        jp = (j + 1) % tube_points
 
-            index_a = i * tube_points + j
-            index_b = ip * tube_points + j
-            index_c = ip * tube_points + jp
-            index_d = i * tube_points + jp
+        index_a = i * tube_points + j
+        index_b = ip * tube_points + j
+        index_c = ip * tube_points + jp
+        index_d = i * tube_points + jp
 
-            indices.append([index_a, index_b, index_d])
-            indices.append([index_b, index_c, index_d])
+        indices.extend(
+            ([index_a, index_b, index_d], [index_b, index_c, index_d])
+        )
     triangles = np.array(indices, dtype=np.uint32)
 
     centers = grid.reshape(grid.shape[0] * grid.shape[1], 3)
@@ -1336,13 +1118,3 @@ def rdp(vertices: npt.NDArray, epsilon: float) -> npt.NDArray:
 
     # When epsilon is 0, avoid removing datapoints
     return vertices
-
-
-if acc_generate_2D_edge_meshes is not None:
-    _generate_2D_edge_meshes = acc_generate_2D_edge_meshes
-else:  # pragma: no cover
-    _generate_2D_edge_meshes = generate_2D_edge_meshes
-
-
-if acc_create_box_from_bounding is not None:
-    create_box_from_bounding = acc_create_box_from_bounding
